requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
mccabe==0.7.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
requests==2.31.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.16
